+++ Gury3d/instance.h
@@ -44,7 +44,7 @@ namespace RBX
 		inline T* findFirstChildOfClass(std::string name)
 		{
 			Instance* child;
-			for (int i = 0; i < getChildren()->size(); i++)
+			for (size_t i = 0; i < getChildren()->size(); i++)
 			{
 				child = getChildren()->at(i);
 				if (child && child->getClassName() == name)
